fix precision

+++ backend/java/src/main/java/com/numetrify/dto/BisectionResponse.java
@@ -3,14 +3,15 @@ package com.numetrify.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 public class BisectionResponse {
     private String message;
-    private List<Double> xVals;
+    private List<BigDecimal> xVals;
     private List<Double> fVals;
-    private List<Double> errors;
+    private List<BigDecimal> errors;
     private List<Integer> iterations;
 }
